+++ be/src/main/java/todolist/dto/card/RequestCardDto.java
@@ -1,8 +1,8 @@
-package todolist.dto;
+package todolist.dto.card;
 
 import lombok.Getter;
 import lombok.Setter;
-import todolist.domain.Card;
+import todolist.domain.card.Card;
 
 @Getter
 @Setter
