+++ be/src/main/java/todolist/dto/card/ResponseCardDto.java
@@ -1,4 +1,4 @@
-package todolist.dto;
+package todolist.dto.card;
 
 public class ResponseCardDto {
