+++ be/src/main/java/todolist/repository/CardMemoryRepository.java
@@ -1,13 +1,11 @@
 package todolist.repository;
 
 import org.springframework.stereotype.Repository;
-import todolist.domain.Card;
-import todolist.dto.RequestCardDto;
+import todolist.domain.card.Card;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 
 @Repository
