+++ be/src/main/java/todolist/service/CardService.java
@@ -2,9 +2,9 @@ package todolist.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import todolist.domain.Card;
-import todolist.dto.RequestCardDto;
-import todolist.dto.ResponseCardDto;
+import todolist.domain.card.Card;
+import todolist.dto.card.RequestCardDto;
+import todolist.dto.card.ResponseCardDto;
 import todolist.repository.TodoRepository;
 
 import java.util.*;
@@ -32,15 +32,6 @@ public class CardService {
     public void updateCard(Long id, RequestCardDto requestCardDto) {
         Card existingCard = repository.findById(id);
         existingCard.update(requestCardDto);
-
-//        if (card.getSection().equals(requestCardDto.getSection())) {
-//            카드 내용 수정 이벤트
-//            eventMaker.create(new Event(Action.UPDATE)
-//        } else {
-//            섹션 이동 이벤트
-//            eventMaker.create(new Event(Action.MOVE)
-//        }
-//        repository.update(existingCard);
     }
 
     public void deleteCard(Long id) {
